Log calnet UID and GIDs, move logging to plugin
